// project/src/types/types-offers.ts
import {TypesCities} from './types-cities';

export type TypesOffers = {
  bedrooms: number;
  city: TypesCities;
  description: string;
  goods: string[];
  host: {
    avatarUrl: string;
    id: number;
    isPro: boolean;
    name: string;
  };
  id: number;
  images: string[];
  isPremium: boolean;
  location: {
    latitude: number;
    longitude: number;
    zoom: number;
  };
  maxAdults: number;
  previewImage: string;
  price: number;
  rating: number;
  title: string;
  type: string;
};


export type OffersType = TypesOffers[];