// project/src/mocks/mocks-offers.ts
import {OffersType} from '../types/types-offers';

const offersMock: OffersType = [
  {
    bedrooms: 2,
    city: {
      location: {
        latitude: 52.370216,
        longitude: 4.895168,
        zoom: 10,
      },
      name: 'Amsterdam',
    },
    description: 'At Hotel Nikko you will find restaurants and lounges that allow you to enjoy the changing colors of gingko trees lining Midosuji Avenue throughout the different seasons.',
    goods: ['Air conditioning', 'Room service', 'Refrigerator', 'Flatscreen TV'],
    host: {
      avatarUrl: 'https://i.pinimg.com/736x/df/4d/0a/df4d0a54c2a2a5c6576299e63ad33cc8--anime-girls-support.jpg',
      id: 125,
      isPro: true,
      name: 'Megumi',
    },
    id: 1,
    images: ['https://dynamic-media-cdn.tripadvisor.com/media/photo-o/09/84/92/26/r.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/13/7d/56/a1/r.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/1c/7e/42/c6/lounge.jpg?w=600&h=-1&s=1'],
    isPremium: true,
    location: {
      latitude: 52.3909553943508,
      longitude: 4.85309666406198,
      zoom: 8,
    },
    maxAdults: 4,
    previewImage: 'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/1c/d4/44/b6/exterior-view.jpg?w=600&h=-1&s=1',
    price: 2000,
    rating: 5,
    title: 'Hotel Nikko',
    type: 'Hotel'
  },

  {
    bedrooms: 1,
    city: {
      location: {
        latitude: 52.370216,
        longitude: 4.895168,
        zoom: 10,
      },
      name: 'Amsterdam',
    },
    description: 'See why so many travelers make Hotel Jakarta Amsterdam their hotel of choice when visiting Amsterdam. Providing an ideal mix of value, comfort and convenience, it offers a romantic setting with an array of amenities designed for travelers like you.',
    goods: ['Private balcony', 'Room service', 'Refrigerator', 'Flatscreen TV'],
    host: {
      avatarUrl: 'https://avatars.mds.yandex.net/i?id=00f27bc9438f7aa3878f11c0aef0404e-7045682-images-thumbs&n=13',
      id: 126,
      isPro: true,
      name: 'Alex',
    },
    id: 2,
    images: ['https://dynamic-media-cdn.tripadvisor.com/media/photo-o/17/ef/21/cf/photo2jpg.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/16/e2/88/5e/hotel-jakarta-amsterdam.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/14/a3/fc/03/the-pool-very-pleasant.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/14/54/2e/cc/photo0jpg.jpg?w=1200&h=-1&s=1'],
    isPremium: true,
    location: {
      latitude: 52.3609553943508,
      longitude: 4.85309666406198,
      zoom: 8,
    },
    maxAdults: 2,
    previewImage: 'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/18/7e/8d/7f/hotel-jakarta-amsterdam.jpg?w=1200&h=-1&s=1',
    price: 1800,
    rating: 4,
    title: 'Hotel Jakarta Amsterdam',
    type: 'Apartment'
  },

  {
    bedrooms: 2,
    city: {
      location: {
        latitude: 52.370216,
        longitude: 4.895168,
        zoom: 10,
      },
      name: 'Amsterdam',
    },
    description: 'The Novotel Paris Les Halles is a 4 star hotel in the centre of Paris with direct Metro and RER train connections to Charles de Gaulle and Orly airports, the major train stations, Disneyland Paris and major Paris attractions including the Pompidou Centre, the Louvre, Notre Dame and more.',
    goods: ['Private balcony', 'Room service', 'Refrigerator', 'Flatscreen TV', 'Extra long beds'],
    host: {
      avatarUrl: 'https://cache3.youla.io/files/images/720_720_out/5d/5e/5d5e7b2b9f359959336d1886.jpg',
      id: 127,
      isPro: true,
      name: 'Marion',
    },
    id: 4,
    images: ['https://dynamic-media-cdn.tripadvisor.com/media/photo-o/12/f6/2a/48/novotel-paris-les-halles.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/19/76/21/cb/novotel-paris-les-halles.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/19/76/21/ae/novotel-paris-les-halles.jpg?w=1200&h=-1&s=1'],
    isPremium: false,
    location: {
      latitude: 52.3909553943508,
      longitude: 4.929309666406198,
      zoom: 8,
    },
    maxAdults: 4,
    previewImage: 'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/26/79/c7/90/novotel-paris-les-halles.jpg?w=1200&h=-1&s=1',
    price: 2300,
    rating: 4,
    title: 'Novotel Les Halles',
    type: 'Apartment'
  },

  {
    bedrooms: 1,
    city: {
      location: {
        latitude: 53.55073,
        longitude: 9.99302,
        zoom: 10,
      },
      name: 'Amsterdam',
    },
    description: 'The Madison Hotel with its 166 rooms is ideally situated between Elbphilharmonie, Speicherstadt and the city, just below the town\'s emblem – St. Michaelis church.',
    goods: ['Room service', 'Refrigerator', 'Flatscreen TV'],
    host: {
      avatarUrl: 'https://st2.depositphotos.com/2668729/5747/v/950/depositphotos_57474431-stock-illustration-avatar-with-a-dollar-sign.jpg',
      id: 128,
      isPro: false,
      name: 'Katherine',
    },
    id: 5,
    images: ['https://dynamic-media-cdn.tripadvisor.com/media/photo-o/12/e4/ba/e4/junior-suite-madison.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/12/e4/bc/57/studio-twin-bed-madison.jpg?w=1200&h=-1&s=1',
      'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/12/e4/af/59/check-in.jpg?w=1200&h=-1&s=1'],
    isPremium: true,
    location: {
      latitude: 52.370216,
      longitude: 4.895168,
      zoom: 8,
    },
    maxAdults: 2,
    previewImage: 'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/12/75/0b/ac/location-madison-hotel.jpg?w=1200&h=-1&s=1',
    price: 1100,
    rating: 3,
    title: 'THE MADISON Hotel',
    type: 'Hotel'
  }
];

export default offersMock;
